requirements.txt
alembic
bcrypt
blinker
click
colorama
Flask
Flask-Bcrypt
Flask-DebugToolbar
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
Flask-RESTful
Flask-Cors
Flask-Session
Flask-Mail
greenlet
itsdangerous
Jinja2
Mako
MarkupSafe
python-dotenv
PyODBC
SQLAlchemy
typing_extensions
Werkzeug
WTForms
pandas
flask_principal